refactor: Replace with new models
