+++ lib/features/session/data/session.dart
@@ -1,21 +1,37 @@
-import 'package:confwebsite2023/features/session/data/talk.dart';
+import 'package:confwebsite2023/features/session/data/speaker.dart';
+import 'package:confwebsite2023/features/session/data/track.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'session.freezed.dart';
 
-@freezed
+part 'session.g.dart';
+
+@Freezed(unionKey: 'type')
 class Session with _$Session {
-  const factory Session.organizer({
+  @JsonSerializable(fieldRename: FieldRename.snake)
+  const factory Session.timeslot({
+    required String uuid,
     required String title,
     required DateTime startsAt,
-    required DateTime endsAt,
-  }) = OrganizerSession;
+    required int lengthMin,
+    required Track track,
+    String? abstract,
+  }) = TimeslotSession;
 
-  const factory Session.talks({
-    required List<Talk> talks,
+  @JsonSerializable(fieldRename: FieldRename.snake)
+  const factory Session.talk({
+    required String uuid,
+    required String url,
+    required String title,
+    required String abstract,
+    required Track track,
     required DateTime startsAt,
-    required DateTime endsAt,
-  }) = TalksSession;
+    required int lengthMin,
+    required Speaker speaker,
+  }) = TalkSession;
 
   const factory Session.lunch() = LunchSession;
+
+  factory Session.fromJson(Map<String, dynamic> json) =>
+      _$SessionFromJson(json);
 }
