+++ lib/features/session/data/speaker.freezed.dart
