--- lib/features/session/data/talk_user.dart
@@ -1,11 +1,0 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'talk_user.freezed.dart';
-
-@freezed
-class TalkUser with _$TalkUser {
-  const factory TalkUser({
-    required String name,
-    required String thumbnailUrl,
-  }) = _TalkUser;
-}
