--- lib/features/session/model/session.dart
@@ -1,36 +1,0 @@
-import 'package:confwebsite2023/features/session/model/speaker.dart';
-import 'package:confwebsite2023/features/session/model/track.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'session.freezed.dart';
-part 'session.g.dart';
-
-@Freezed(unionKey: 'type')
-class Session with _$Session {
-  @JsonSerializable(fieldRename: FieldRename.snake)
-  const factory Session.timeslot({
-    required String uuid,
-    required String title,
-    required DateTime startsAt,
-    required int lengthMin,
-    required Track track,
-    String? abstract,
-  }) = TimeslotSession;
-
-  @JsonSerializable(fieldRename: FieldRename.snake)
-  const factory Session.talk({
-    required String uuid,
-    required String url,
-    required String title,
-    required String abstract,
-    required Track track,
-    required DateTime startsAt,
-    required int lengthMin,
-    required Speaker speaker,
-  }) = TalkSession;
-
-  const factory Session.lunch() = LunchSession;
-
-  factory Session.fromJson(Map<String, dynamic> json) =>
-      _$SessionFromJson(json);
-}
